Make call to job queue in DomainThread not be blocking
The queue associated with the threads blocks waiting for another item, so once they're all gone each thread will be unable to progress, and the call to kill them later on does nothing. Making the call be a non-blocking one and catching the exception means the threads do actually exit.